Simplified the WebSocketServer API.
Incoming HTTP requests are now parsed prior to delivering the connection to
the event queue. Event handlers are now only responsible for sending a response
or upgrading to a WebSocket.

Improved the EchoServer example with better stats and an randomized URL that
can be used to shut down the server from the browser.