WebSocket: Added a service task.
The implementation was previously relying on the application task that iterates
over events from the WebSocket object to drive the process of connecting, performing
the opening handshake, and processing incoming frames. This is problematic for
the common use case of sending an initial message to the other endpoint prior to
entering the event loop. In that scenario, the async send() never completes because
doing so requires the socket to be connected, which in turn requires event iteration
to have started. Obviously, the workaround is to send the initial message upon
receipt of the `open` event, but send() is explicitly documented to allow "early"
messages that will be delivered once the connection is established.

This commit adds a new service task internal to each WebSocket that is started
when 1) any of the APIs that send frames are called, or 2) the next event is
requested. The service task handles connecting, the opening handshake,
and incoming frame processing, dropping events into a queue that the
application can drain at its leisure.