WebSocket: Revisiting the service task.
This commit aims to bring back the benefits of the service task without the drawbacks.
The task iterating events still drives the servicing of the socket, but the first attempt
to send a message while in the `initialized` ready state now performs the connect
and opening handshake, instead of leaving it to the event iterator. This ensures that
every send will complete, even if performed by the task that later begins consuming
events.

The behavior of calling `close(::)` while in the `connecting` state has changed.
Instead of immediately forcing the socket into the `closed` state, it waits for the
opening handshake to complete before initiating a graceful close (assuming the
handshake is successful). This change ensures that a series of sends followed
by a close will end up delivering all of those frames to the remote endpoint.

Additionally, the output framer was slightly simplified to reflect how it's actually
used. One notable difference is that it always succeeds. Previously it was allowed
to fail if a `ping` or `pong` frame contained too much data. Now it simply truncates
the data in those cases.