Initial compression work plus some optimizations.
Connection: Added the ability to send an array of Data structs using NWConnection's
"batch" feature.

Framing: Frames are now encoded into two Data structs: the header and payload.
The advantage of this is that when sending unmasked frames, we no longer need to
build a concatenated header+payload buffer. Since the server sends unmasked
frames, and also tends to be the endpoint sending large frames, this should be a
good optimization. The mask function was also sped up by working with the
underlying UnsafeMutableBufferPointer directly.