Close handshake improvements.
The close code and reason reported by the `close` event are now those
received from the remote endpoint, even if the local endpoint initiated the
closing handshake. This is more in line with how other implementations
handle the event.

Additionally, it is now possible to close the WebSocket with a `nil` close
code, which sends a `close` frame with no payload at all.

The low-level Connection class now ensures that frames composed of
multiple buffers are now delivered to the network stack as a single unit.
Previously, two separate TCP/IP frames were being generated for each
WebSocket frame (header and payload).